+++ apm-agent-core/src/main/java/co/elastic/apm/agent/bci/ElasticApmAgent.java
@@ -497,7 +497,7 @@ public class ElasticApmAgent {
             }
         }
         if (!(classLoader instanceof ExternalPluginClassLoader) && adviceClassName.startsWith("co.elastic.apm.agent.") && adviceClassName.split("\\.").length > 6) {
-            throw new IllegalStateException("Indy-dispatched advice class must be at the root of the instrumentation plugin.");
+            throw new IllegalStateException("Indy-dispatched advice class must be at the root of the instrumentation plugin: " + adviceClassName);
         }
     }
