+++ apm-agent-plugins/apm-rabbitmq/apm-rabbitmq-plugin/src/main/resources/META-INF/services/co.elastic.apm.agent.sdk.ElasticApmInstrumentation
@@ -1,5 +1,3 @@
 co.elastic.apm.agent.rabbitmq.ChannelInstrumentation$BasicPublish
 co.elastic.apm.agent.rabbitmq.ChannelInstrumentation$BasicConsume
 co.elastic.apm.agent.rabbitmq.ChannelInstrumentation$BasicGet
-co.elastic.apm.agent.rabbitmq.SetMessageListenerInstrumentation
-co.elastic.apm.agent.rabbitmq.SpringAmqpMessageListenerInstrumentation
