+++ apm-agent-plugins/apm-rabbitmq/apm-rabbitmq-spring/src/main/java/co/elastic/apm/agent/rabbitmq/MessageListenerHelper.java
@@ -11,9 +11,9 @@
  * the Apache License, Version 2.0 (the "License"); you may
  * not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *   http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
@@ -22,17 +22,34 @@
  * under the License.
  * #L%
  */
-package co.elastic.apm.agent.rabbitmq.config;
+package co.elastic.apm.agent.rabbitmq;
 
+import org.springframework.amqp.core.Message;
 import org.springframework.amqp.core.MessageListener;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
 
-@Configuration
-public class LambdaMessageListenerConfiguration extends DefaultBindingSpringConfiguration {
+import javax.annotation.Nullable;
 
-    @Bean
-    public MessageListener messageListener() {
-        return message -> testSpan();
+public class MessageListenerHelper {
+
+    @Nullable
+    public MessageListener wrapLambda(@Nullable MessageListener listener) {
+        if (listener != null && listener.getClass().getName().contains("/")) {
+            return new MessageListenerWrapper(listener);
+        }
+        return listener;
+    }
+
+    public static class MessageListenerWrapper implements MessageListener {
+
+        private final MessageListener delegate;
+
+        public MessageListenerWrapper(MessageListener delegate) {
+            this.delegate = delegate;
+        }
+
+        @Override
+        public void onMessage(Message message) {
+            delegate.onMessage(message);
+        }
     }
 }
