+++ apm-agent-plugins/apm-rabbitmq/apm-rabbitmq-spring/src/main/resources/META-INF/services/co.elastic.apm.agent.sdk.ElasticApmInstrumentation
@@ -1,0 +1,2 @@
+co.elastic.apm.agent.rabbitmq.SetMessageListenerInstrumentation
+co.elastic.apm.agent.rabbitmq.SpringAmqpMessageListenerInstrumentation
