+++ apm-agent-plugins/apm-rabbitmq/apm-rabbitmq-spring/src/test/java/co/elastic/apm/agent/rabbitmq/config/DirectMessageListenerContainerConfiguration.java
@@ -24,30 +24,20 @@
  */
 package co.elastic.apm.agent.rabbitmq.config;
 
-import org.springframework.amqp.rabbit.annotation.EnableRabbit;
-import org.springframework.amqp.rabbit.annotation.RabbitListener;
-import org.springframework.amqp.rabbit.config.SimpleRabbitListenerContainerFactory;
+import co.elastic.apm.agent.rabbitmq.TestConstants;
 import org.springframework.amqp.rabbit.connection.ConnectionFactory;
-import org.springframework.amqp.support.converter.Jackson2JsonMessageConverter;
+import org.springframework.amqp.rabbit.listener.DirectMessageListenerContainer;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
-import static co.elastic.apm.agent.rabbitmq.TestConstants.QUEUE_NAME;
-
 @Configuration
-@EnableRabbit
-public class RabbitListenerConfiguration extends DefaultBindingSpringConfiguration {
-
-    @RabbitListener(queues = QUEUE_NAME)
-    public void processMessage(String message) {
-        testSpan();
-    }
+public class DirectMessageListenerContainerConfiguration extends MessageListenerConfiguration {
 
     @Bean
-    public SimpleRabbitListenerContainerFactory rabbitListenerContainerFactory(ConnectionFactory connectionFactory) {
-        SimpleRabbitListenerContainerFactory factory = new SimpleRabbitListenerContainerFactory();
-        factory.setConnectionFactory(connectionFactory);
-        factory.setMessageConverter(new Jackson2JsonMessageConverter());
-        return factory;
+    DirectMessageListenerContainer container(ConnectionFactory connectionFactory) {
+        DirectMessageListenerContainer directMessageListenerContainer = new DirectMessageListenerContainer(connectionFactory);
+        directMessageListenerContainer.setQueueNames(TestConstants.QUEUE_NAME);
+        directMessageListenerContainer.setMessageListener(messageListener());
+        return directMessageListenerContainer;
     }
 }
