+++ apm-agent-plugins/apm-rabbitmq/pom.xml
@@ -21,7 +21,7 @@
         <module>apm-rabbitmq-plugin</module>
         <module>apm-rabbitmq-test-3</module>
         <module>apm-rabbitmq-test-4</module>
-        <module>apm-rabbitmq-spring-test</module>
+        <module>apm-rabbitmq-spring</module>
         <!-- RabbitMQ 5.x is already tested through plugin module, thus we don't need a separate module to cover it -->
     </modules>
